Aggregation of records based on same-as relationships
Adding a LookupService implementation which delegates to another service
and post-processes the results to sort them to co-locate candidates that
have a same-as relationship.

There are also two new lookup configurations aggregated-federated-lookup
and aggregated-default-lookup for rds-global which use this service.

This change now requires a 4.1-based platform.war as compile dependency.