Implement ability to search by identifier
Searching by identifier (local name) of a resource (e.g. W42 for Douglas
Adams in Wikidata) was not previously possible as the identifiers
themselves are not indexed.
This change adds a search pattern where the search token is used to
construct a possible IRI from the datasets's main namespace(s) and if a
rdf:type statement for this resource exists in the target dataset, then
it is included as a result candidate.