Using different predicate to express same-as relations
When materializing same-as relationships we now use the predicate
http://schema.swissartresearch.net/ontology/rds#related instead of
https://static.swissartresearch.net/sameAs to express connections
between entities. This is to hint at a less-string relationship between
entities.
The relevant places to adjust this are:
- data-preparation/groupingFeatureAdjustment/DockerImage/transform.js
for the materialization process
- The lookup service configurations fetching entity references
(rds-global/config/repositories/aggregated-default-lookup.ttl,
rds-global/config/repositories/aggregated-federated-lookup.ttl)

The materialization scripts now report errors when fetching data from
the database.

Additionally, this change contains some additional cleanup updates for
the RDS apps.