RDS-210 Apps as volumes
Apps are typically deployed to /apps which is converted into a Docker
volume on first start and any existing content from /apps is copied into
the volume only once when creating the volume. For baked-in apps like
the set of published RDS-L apps that would mean that any updates in
newer container images were never re-deployed.

The set of apps that make up the RDS-L deployment are now published in
the /rdsapps folder within the container instead and activated via
environment variables MP_APP_<N> pointing to the respective app.

Any additional apps deployed to /apps behave like before and are visible
in addition to the ones in /rdsapps.

As the RDS-L apps are deployed as folder the start time is actually a
bit faster as well as they do not need to be unzipped on first start.